Moved bot actions into concern which will be shared by groups and projects
